Organized css files and added basic functionality to sequencer
The sequencer has issues with needing to mute notes while recording a sequence. If a note is played while the sequencer is also recording then the note never gets released. Another issue is that the sequencer keeps getting reset after toggling the record button. It should only reset the sequence when toggling the record button on.