Created functioning AmpEnv that can be updated.
Created functioning AmpEnv that can be updated with circle sliders or anything really. Annoying fix and is probably not perfect but it will do for now.
a useState hook was made initialized with a new Tone.Envelope object. to update the AmpEnv the values in the AmpEnvState hook are updated and after any update the AmpEnv hook is updated to create a new Tone.Envelope object with the new values inside of ampEnvState.
Also fixed issue of having multiple functions for changing volume. This was necessary as the envelope will have 4 circle sliders and I do not want unnecessary functions. Especially if there will be another filter envelope.