Updated Play and Release functions
Updated Release And Play functions.
Instead of midi playing notes itself it calls the new renamed handleReleaseNote and handlePlayNote function. It now gets passed a note to play instead of an event.
Keyboard click and release were updated to destructure the event before passing it to the play and release functions.