Set up temporary preset backend
Created a temporary backend for storing all the preset data.

There was an issue with the envelope updating 1 behind what it needed to. (selecting a preset would update everything except the ampenvstate would be set to the previous values. so if you selected preset 2 it would have the last presets amp values. if you then selected preset 3 then it would have preset 2's amp values)
To fix it I just added a use effect to update the ampEnvelope every time the state changed.